Break out keyboard events + rename files
