+++ lib/handle-click.js
@@ -1,6 +1,6 @@
 import { colorLuminance } from './utils';
-import { getModal } from './swal-dom';
-import { hasClass, isDescendant } from './dom-manipulation';
+import { getModal } from './handle-swal-dom';
+import { hasClass, isDescendant } from './handle-dom';
 
 
 /*
