+++ lib/handle-dom.js
