+++ lib/handle-swal-dom.js
@@ -1,5 +1,5 @@
 import { hexToRgb } from './utils';
-import { removeClass, getTopMargin, fadeIn, show, addClass } from './dom-manipulation';
+import { removeClass, getTopMargin, fadeIn, show, addClass } from './handle-dom';
 import defaultParams from './default-params';
 
 var modalClass   = '.sweet-alert';
