+++ lib/set-params.js
@@ -8,13 +8,14 @@ import {
   getModal,
   getInput,
   setFocusStyle
-} from './swal-dom';
+} from './handle-swal-dom';
 
 import {
   hasClass, addClass, removeClass, 
   escapeHtml, 
   _show, show, _hide, hide
-} from './dom-manipulation';
+} from './handle-dom';
+
 
 /*
  * Set type, text and actions on modal
@@ -22,56 +23,65 @@ import {
 var setParameters = function(params) {
   var modal = getModal();
 
-  var $title = modal.querySelector('h2'),
-      $text = modal.querySelector('p'),
-      $cancelBtn = modal.querySelector('button.cancel'),
-      $confirmBtn = modal.querySelector('button.confirm');
+  var $title = modal.querySelector('h2');
+  var $text = modal.querySelector('p');
+  var $cancelBtn = modal.querySelector('button.cancel');
+  var $confirmBtn = modal.querySelector('button.confirm');
 
-  // Title
+  /*
+   * Title
+   */
   $title.innerHTML = params.html ? params.title : escapeHtml(params.title).split('\n').join('<br>');
 
-  // Text
+  /*
+   * Text
+   */
   $text.innerHTML = params.html ? params.text : escapeHtml(params.text || '').split('\n').join('<br>');
+  if (params.text) show($text);
 
-  if (params.text) {
-    show($text);
-  }
-
-  //Custom Class
+  /*
+   * Custom class
+   */
   if (params.customClass) {
     addClass(modal, params.customClass);
     modal.setAttribute('data-custom-class', params.customClass);
   } else {
     // Find previously set classes and remove them
-    var customClass = modal.getAttribute('data-custom-class');
+    let customClass = modal.getAttribute('data-custom-class');
     removeClass(modal, customClass);
     modal.setAttribute('data-custom-class', '');
   }
 
-  // Icon
+  /*
+   * Icon
+   */
   hide(modal.querySelectorAll('.sa-icon'));
+
   if (params.type && !isIE8()) {
-    var validType = false;
-    for (var i = 0; i < alertTypes.length; i++) {
+
+    let validType = false;
+
+    for (let i = 0; i < alertTypes.length; i++) {
       if (params.type === alertTypes[i]) {
         validType = true;
         break;
       }
     }
+
     if (!validType) {
       logStr('Unknown alert type: ' + params.type);
       return false;
     }
 
-    var typesWithIcons = ['success', 'error', 'warning', 'info'];
-    var $icon;
+    let typesWithIcons = ['success', 'error', 'warning', 'info'];
+    let $icon;
 
     if (typesWithIcons.indexOf(params.type) !== -1) {
       $icon = modal.querySelector('.sa-icon.' + 'sa-' + params.type);
       show($icon);
     }
 
-    var $input = getInput();
+    let $input = getInput();
 
     // Animate icon
     switch (params.type) {
@@ -106,20 +116,22 @@ var setParameters = function(params) {
     }
   }
 
-  // Custom image
+  /*
+   * Custom image
+   */
   if (params.imageUrl) {
-    var $customIcon = modal.querySelector('.sa-icon.sa-custom');
+    let $customIcon = modal.querySelector('.sa-icon.sa-custom');
 
     $customIcon.style.backgroundImage = 'url(' + params.imageUrl + ')';
     show($customIcon);
 
-    var _imgWidth = 80,
-        _imgHeight = 80;
+    let _imgWidth = 80;
+    let _imgHeight = 80;
 
     if (params.imageSize) {
-      var dimensions = params.imageSize.toString().split('x');
-      var imgWidth = dimensions[0];
-      var imgHeight = dimensions[1];
+      let dimensions = params.imageSize.toString().split('x');
+      let imgWidth = dimensions[0];
+      let imgHeight = dimensions[1];
 
       if (!imgWidth || !imgHeight) {
         logStr('Parameter imageSize expects value with format WIDTHxHEIGHT, got ' + params.imageSize);
@@ -128,10 +140,13 @@ var setParameters = function(params) {
         _imgHeight = imgHeight;
       }
     }
+
     $customIcon.setAttribute('style', $customIcon.getAttribute('style') + 'width:' + _imgWidth + 'px; height:' + _imgHeight + 'px');
   }
 
-  // Show cancel button?
+  /*
+   * Show cancel button?
+   */
   modal.setAttribute('data-has-cancel-button', params.showCancelButton);
   if (params.showCancelButton) {
     $cancelBtn.style.display = 'inline-block';
@@ -139,7 +154,9 @@ var setParameters = function(params) {
     hide($cancelBtn);
   }
 
-  // Show confirm button?
+  /*
+   * Show confirm button?
+   */
   modal.setAttribute('data-has-confirm-button', params.showConfirmButton);
   if (params.showConfirmButton) {
     $confirmBtn.style.display = 'inline-block';
@@ -147,7 +164,9 @@ var setParameters = function(params) {
     hide($confirmBtn);
   }
 
-  // Edit text on cancel and confirm buttons
+  /*
+   * Custom text on cancel/confirm buttons
+   */
   if (params.cancelButtonText) {
     $cancelBtn.innerHTML = escapeHtml(params.cancelButtonText);
   }
@@ -155,6 +174,9 @@ var setParameters = function(params) {
     $confirmBtn.innerHTML = escapeHtml(params.confirmButtonText);
   }
 
+  /*
+   * Custom color on confirm button
+   */
   if (params.confirmButtonColor) {
     // Set confirm button to selected background color
     $confirmBtn.style.backgroundColor = params.confirmButtonColor;
@@ -163,15 +185,21 @@ var setParameters = function(params) {
     setFocusStyle($confirmBtn, params.confirmButtonColor);
   }
 
-  // Allow outside click?
+  /*
+   * Allow outside click
+   */
   modal.setAttribute('data-allow-outside-click', params.allowOutsideClick);
 
-  // Done-function
+  /*
+   * Callback function
+   */
   var hasDoneFunction = params.doneFunction ? true : false;
   modal.setAttribute('data-has-done-function', hasDoneFunction);
 
+  /*
+   * Animation
+   */
   if (!params.animation) {
-    // No animation
     modal.setAttribute('data-animation', 'none');
   } else if (typeof params.animation === 'string') {
     modal.setAttribute('data-animation', params.animation); // Custom animation
@@ -179,7 +207,9 @@ var setParameters = function(params) {
     modal.setAttribute('data-animation', 'pop');
   }
 
-  // Close timer
+  /*
+   * Timer
+   */
   modal.setAttribute('data-timer', params.timer);
 }
