+++ lib/sweet-alert.js
@@ -16,7 +16,7 @@ import {
   fadeIn, fadeOut,
   fireClick,
   stopEventPropagation
-} from './dom-manipulation';
+} from './handle-dom';
 
 /*
  * Handy utilities
@@ -41,14 +41,12 @@ import {
   openModal,
   resetInput,
   fixVerticalPosition
-} from './swal-dom';
+} from './handle-swal-dom';
 
 
-import { 
-  handleButton, 
-  handleConfirm, 
-  handleCancel 
-} from './button-click';
+// Handle button events and keyboard events
+import { handleButton, handleConfirm, handleCancel } from './handle-click';
+import handleKeyDown from './handle-key';
 
 
 // Default values
@@ -65,6 +63,7 @@ var lastFocusedButton;
 
 /*
  * Global sweetAlert function
+ * (this is what the user calls)
  */
 var sweetAlert, swal;
 
@@ -81,7 +80,7 @@ sweetAlert = swal = function() {
    */
   function argumentOrDefault(key) {
     var args = customizations;
-    return args[key] || defaultParams[key];
+    return (args[key] === undefined) ?  defaultParams[key] : args[key];
   }
 
   if (customizations === undefined) {
@@ -100,7 +99,7 @@ sweetAlert = swal = function() {
       params.type  = arguments[2] || '';
       break;
 
-    // Ex: swal({title:"Hello", text: "Just testing", type: "info"});
+    // Ex: swal({ title:"Hello", text: "Just testing", type: "info" });
     case 'object':
       if (customizations.title === undefined) {
         logStr('Missing "title" argument!');
@@ -153,84 +152,16 @@ sweetAlert = swal = function() {
   // Clicking outside the modal dismisses it (if allowed by user)
   getOverlay().onclick = onButtonEvent;
 
-
-  // Keyboard interactions
-  var $okButton     = modal.querySelector('button.confirm');
-  var $cancelButton = modal.querySelector('button.cancel');
-  var $modalButtons = modal.querySelectorAll('button[tabindex]');
-
-  function handleKeyDown(event) {
-    var e = event || window.event;
-    var keyCode = e.keyCode || e.which;
-
-    if ([9, 13, 32, 27].indexOf(keyCode) === -1) {
-      // Don't do work on keys we don't care about.
-      return;
-    }
-
-    var $targetElement = e.target || e.srcElement;
-
-    var btnIndex = -1; // Find the button - note, this is a nodelist, not an array.
-    for (var i = 0; i < $modalButtons.length; i++) {
-      if ($targetElement === $modalButtons[i]) {
-        btnIndex = i;
-        break;
-      }
-    }
-
-    if (keyCode === 9) {
-      // TAB
-      if (btnIndex === -1) {
-        // No button focused. Jump to the confirm button.
-        $targetElement = $okButton;
-      } else {
-        // Cycle to the next button
-        if (btnIndex === $modalButtons.length - 1) {
-          $targetElement = $modalButtons[0];
-        } else {
-          $targetElement = $modalButtons[btnIndex + 1];
-        }
-      }
-
-      stopEventPropagation(e);
-      $targetElement.focus();
-
-      if (params.confirmButtonColor) {
-        setFocusStyle($targetElement, params.confirmButtonColor);
-      }
-    } else {
-      if (keyCode === 13) {
-        if ($targetElement.tagName === 'INPUT') {
-          $targetElement = $okButton;
-          $okButton.focus();
-        }
-
-        if (btnIndex === -1) {
-          // ENTER/SPACE clicked outside of a button.
-          $targetElement = $okButton;
-        } else {
-          // Do nothing - let the browser handle it.
-          $targetElement = undefined;
-        }
-      } else if (keyCode === 27 && params.allowEscapeKey === true) {
-        $targetElement = $cancelButton;
-        fireClick($targetElement, e);
-      } else {
-        // Fallback - let the browser handle it.
-        $targetElement = undefined;
-      }
-    }
-  }
-
   previousWindowKeyDown = window.onkeydown;
 
-  window.onkeydown = handleKeyDown;
+  var onKeyEvent = (e) => handleKeyDown(e, params, modal);
+  window.onkeydown = onKeyEvent;
 
   window.onfocus = function () {
     // When the user has focused away and focused back from the whole window.
     setTimeout(function () {
-      // Put in a timeout to jump out of the event sequence. Calling focus() in the event
-      // sequence confuses things.
+      // Put in a timeout to jump out of the event sequence.
+      // Calling focus() in the event sequence confuses things.
       if (lastFocusedButton !== undefined) {
         lastFocusedButton.focus();
         lastFocusedButton = undefined;
@@ -238,6 +169,7 @@ sweetAlert = swal = function() {
     }, 0);
   };
 };
+
 
 
 /*
@@ -268,8 +200,9 @@ sweetAlert.close = swal.close = function() {
   addClass(modal, 'hideSweetAlert');
   removeClass(modal, 'visible');
 
-  // Reset icon animations
-
+  /*
+   * Reset icon animations
+   */
   var $successIcon = modal.querySelector('.sa-icon.sa-success');
   removeClass($successIcon, 'animate');
   removeClass($successIcon.querySelector('.sa-tip'), 'animateSuccessTip');
@@ -284,6 +217,7 @@ sweetAlert.close = swal.close = function() {
   removeClass($warningIcon.querySelector('.sa-body'), 'pulseWarningIns');
   removeClass($warningIcon.querySelector('.sa-dot'), 'pulseWarningIns');
 
+  // Make page scrollable again
   removeClass(document.body, 'stop-scrolling');
 
   // Reset the page to its previous state
@@ -293,6 +227,8 @@ sweetAlert.close = swal.close = function() {
   }
   lastFocusedButton = undefined;
   clearTimeout(modal.timeout);
+
+  return true;
 };
